refactor(WinMoney): 불필요한 클래스 WinMoney 삭제
