+++ src/main/kotlin/lotto/model/LottoProfitRateCalculator.kt
@@ -1,10 +1,11 @@
 package lotto.model
 
-import lotto.entity.*
+import lotto.entity.Money
+import lotto.entity.ProfitRate
+import lotto.entity.PurchaseMoney
 
 class LottoProfitRateCalculator : ProfitRateCalculator {
-    override fun calculate(purchaseMoney: PurchaseMoney, winMoney: WinMoney): ProfitRate {
+    override fun calculate(purchaseMoney: PurchaseMoney, winMoney: Money): ProfitRate {
         return ProfitRate(winMoney.value / purchaseMoney.value.toFloat())
     }
-
 }
