+++ src/main/kotlin/lotto/model/ProfitRateCalculator.kt
@@ -1,9 +1,9 @@
 package lotto.model
 
+import lotto.entity.Money
 import lotto.entity.ProfitRate
 import lotto.entity.PurchaseMoney
-import lotto.entity.WinMoney
 
 interface ProfitRateCalculator {
-    fun calculate(purchaseMoney: PurchaseMoney, winMoney: WinMoney): ProfitRate
+    fun calculate(purchaseMoney: PurchaseMoney, winMoney: Money): ProfitRate
 }
